+++ src/dialogs/save_file_dialog.rs
@@ -136,18 +136,18 @@ impl<SR: Clone + Debug + PartialEq, SN: Clone + Debug + PartialEq> SaveFileDialo
 where
     SR: FileDialogResults<SR>,
 {
-    fn render(&self, graphics: &mut Graphics, mouse_xy: Coord) {
+    fn render(&self, graphics: &mut Graphics, mouse: &MouseData) {
         self.background.render(graphics);
-        self.name_field.render(graphics, mouse_xy);
-        self.current_dir_field.render(graphics, mouse_xy);
-        self.dir_panel.render(graphics, mouse_xy);
-        self.save.render(graphics, mouse_xy);
-        self.home.render(graphics, mouse_xy);
-        self.downloads.render(graphics, mouse_xy);
-        self.docs.render(graphics, mouse_xy);
-        self.home.render(graphics, mouse_xy);
-        self.load.render(graphics, mouse_xy);
-        self.cancel.render(graphics, mouse_xy);
+        self.name_field.render(graphics, mouse);
+        self.current_dir_field.render(graphics, mouse);
+        self.dir_panel.render(graphics, mouse);
+        self.save.render(graphics, mouse);
+        self.home.render(graphics, mouse);
+        self.downloads.render(graphics, mouse);
+        self.docs.render(graphics, mouse);
+        self.home.render(graphics, mouse);
+        self.load.render(graphics, mouse);
+        self.cancel.render(graphics, mouse);
     }
 
     fn update(&mut self, timing: &Timing) -> SceneUpdateResult<SR, SN> {
@@ -163,30 +163,42 @@ where
     SR: FileDialogResults<SR>,
 {
     #[cfg(any(feature = "controller", feature = "controller_xinput"))]
-    fn render(&self, graphics: &mut Graphics, mouse_xy: Coord, _: &[KeyCode], _: &GameController) {
-        self.render(graphics, mouse_xy)
+    fn render(
+        &self,
+        graphics: &mut Graphics,
+        mouse: &MouseData,
+        _: &[KeyCode],
+        _: &GameController,
+    ) {
+        self.render(graphics, mouse)
     }
 
     #[cfg(not(any(feature = "controller", feature = "controller_xinput")))]
-    fn render(&self, graphics: &mut Graphics, mouse_xy: Coord, _: &[KeyCode]) {
-        self.render(graphics, mouse_xy)
+    fn render(&self, graphics: &mut Graphics, mouse: &MouseData, _: &[KeyCode]) {
+        self.render(graphics, mouse)
     }
 
-    fn on_key_up(&mut self, key: KeyCode, _: Coord, held_keys: &[KeyCode]) {
+    fn on_key_up(&mut self, key: KeyCode, _: &MouseData, held_keys: &[KeyCode]) {
         self.name_field.on_key_press(key, held_keys);
         self.current_dir_field.on_key_press(key, held_keys);
     }
 
-    fn on_mouse_up(&mut self, xy: Coord, button: MouseButton, _: &[KeyCode]) {
+    fn on_mouse_click(
+        &mut self,
+        down_at: Coord,
+        mouse: &MouseData,
+        button: MouseButton,
+        _: &[KeyCode],
+    ) {
         if button != MouseButton::Left {
             return;
         }
-        if self.cancel.on_mouse_click(xy) {
+        if self.cancel.on_mouse_click(down_at, mouse.xy) {
             self.result = SceneUpdateResult::Pop(None);
         }
-        self.current_dir_field.on_mouse_click(xy);
-        self.name_field.on_mouse_click(xy);
-        if let Some(result) = self.dir_panel.on_mouse_click(xy) {
+        self.current_dir_field.on_mouse_click(down_at, mouse.xy);
+        self.name_field.on_mouse_click(down_at, mouse.xy);
+        if let Some(result) = self.dir_panel.on_mouse_click(down_at, mouse.xy) {
             if result.is_file {
                 let filename = PathBuf::from(result.path)
                     .file_name()
@@ -199,7 +211,7 @@ where
                 self.current_dir_field.set_content(&result.path);
             }
         }
-        if self.downloads.on_mouse_click(xy) {
+        if self.downloads.on_mouse_click(down_at, mouse.xy) {
             self.current_dir_field.set_content(
                 &UserDirs::new()
                     .unwrap()
@@ -209,7 +221,7 @@ where
             );
             self.dir_panel.set_dir(self.current_dir_field.content());
         }
-        if self.docs.on_mouse_click(xy) {
+        if self.docs.on_mouse_click(down_at, mouse.xy) {
             self.current_dir_field.set_content(
                 &UserDirs::new()
                     .unwrap()
@@ -219,12 +231,12 @@ where
             );
             self.dir_panel.set_dir(self.current_dir_field.content());
         }
-        if self.home.on_mouse_click(xy) {
+        if self.home.on_mouse_click(down_at, mouse.xy) {
             self.current_dir_field
                 .set_content(&UserDirs::new().unwrap().home_dir().to_string_lossy());
             self.dir_panel.set_dir(self.current_dir_field.content());
         }
-        if self.save.on_mouse_click(xy) && !self.name_field.content().is_empty() {
+        if self.save.on_mouse_click(down_at, mouse.xy) && !self.name_field.content().is_empty() {
             let mut path = PathBuf::from(self.current_dir_field.content());
             path.push(self.name_field.content());
             if let Some(ext) = &self.expected_ext {
@@ -238,15 +250,15 @@ where
         }
     }
 
-    fn on_scroll(&mut self, xy: Coord, _: isize, y_diff: isize, _: &[KeyCode]) {
-        self.dir_panel.on_scroll(xy, y_diff);
+    fn on_scroll(&mut self, mouse: &MouseData, _: isize, y_diff: isize, _: &[KeyCode]) {
+        self.dir_panel.on_scroll(mouse.xy, y_diff);
     }
 
     #[cfg(any(feature = "controller", feature = "controller_xinput"))]
     fn update(
         &mut self,
         timing: &Timing,
-        _: Coord,
+        _: &MouseData,
         _: &[KeyCode],
         _: &GameController,
     ) -> SceneUpdateResult<SR, SN> {
@@ -254,7 +266,12 @@ where
     }
 
     #[cfg(not(any(feature = "controller", feature = "controller_xinput")))]
-    fn update(&mut self, timing: &Timing, _: Coord, _: &[KeyCode]) -> SceneUpdateResult<SR, SN> {
+    fn update(
+        &mut self,
+        timing: &Timing,
+        _: &MouseData,
+        _: &[KeyCode],
+    ) -> SceneUpdateResult<SR, SN> {
         self.update(timing)
     }