+++ src/ui/button.rs
@@ -4,6 +4,7 @@ use crate::buffer_graphics_lib::text::format::Positioning::Center;
 use crate::buffer_graphics_lib::text::pos::TextPos;
 use crate::buffer_graphics_lib::text::wrapping::WrappingStrategy;
 use crate::buffer_graphics_lib::text::Text;
+use crate::scenes::MouseData;
 use crate::ui::styles::ButtonStyle;
 use crate::ui::{ElementState, UiElement};
 use simple_game_utils::prelude::Timing;
@@ -83,9 +84,9 @@ impl Button {
     }
 
     #[must_use]
-    pub fn on_mouse_click(&mut self, xy: Coord) -> bool {
+    pub fn on_mouse_click(&mut self, down: Coord, up: Coord) -> bool {
         if self.state != ElementState::Disabled {
-            self.bounds.contains(xy)
+            self.bounds.contains(down) && self.bounds.contains(up)
         } else {
             false
         }
@@ -106,9 +107,9 @@ impl UiElement for Button {
         &self.bounds
     }
 
-    fn render(&self, graphics: &mut Graphics, mouse_xy: Coord) {
+    fn render(&self, graphics: &mut Graphics, mouse: &MouseData) {
         let (error, disabled) = self.state.get_err_dis();
-        let hovering = self.bounds.contains(mouse_xy);
+        let hovering = self.bounds.contains(mouse.xy);
         if let Some(color) = self.style.shadow.get(hovering, error, disabled) {
             self.shadow.with_color(color).render(graphics);
         }