+++ src/ui/icon_button.rs
@@ -91,9 +91,9 @@ impl IconButton {
 
 impl IconButton {
     #[must_use]
-    pub fn on_mouse_click(&mut self, xy: Coord) -> bool {
+    pub fn on_mouse_click(&mut self, down: Coord, up: Coord) -> bool {
         if self.state != ElementState::Disabled {
-            self.bounds.contains(xy)
+            self.bounds.contains(down) && self.bounds.contains(up)
         } else {
             false
         }
@@ -121,9 +121,9 @@ impl UiElement for IconButton {
         &self.bounds
     }
 
-    fn render(&self, graphics: &mut Graphics, mouse_xy: Coord) {
+    fn render(&self, graphics: &mut Graphics, mouse: &MouseData) {
         let (error, disabled) = self.state.get_err_dis();
-        let hovering = self.bounds.contains(mouse_xy);
+        let hovering = self.bounds.contains(mouse.xy);
         if let Some(color) = self.style.shadow.get(hovering, error, disabled) {
             self.shadow.with_color(color).render(graphics);
         }
@@ -132,7 +132,7 @@ impl UiElement for IconButton {
         }
         graphics.draw_indexed_image(self.icon_xy, &self.icon);
         if !disabled && hovering {
-            self.tooltip.render(graphics, mouse_xy);
+            self.tooltip.render(graphics, mouse);
         }
     }